refactor(orchestrator): extract recall-section budgeting/assembly coordinator (seam 13, #1526)
Extract the recall section budgeting and assembly subsystem (section
enablement lookups, per-section maxChars/number resolution, section
bucket appends, budget-aware truncation, ordered assembly,
LastRecallBudgetSummary, source collection) into a new
RecallSectionCoordinator using the lazy-getter coordinator pattern.

Behavior-preserving: the orchestrator keeps thin delegating methods so
all call sites in recallInternal and the consolidation loop continue to
work unchanged.

- orchestrator.ts: 18679 → 18506 LOC (−173)
- New: orchestration/recall-section-coordinator.ts (329 LOC)
- Repointed config-recall-pipeline.test.ts to wire the coordinator
- Ratchet baseline updated (LOC may only decrease)
